[REFACTOR] move SocketIOEvent type -> constant
