+++ src/main/java/com/cone/cone/external/socket/ChatFacadeImpl.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.cone.cone.external.socket.type.SocketIOEvent.MESSAGE;
+import static com.cone.cone.external.socket.constant.SocketIOEvent.MESSAGE;
 
 @Component
 @RequiredArgsConstructor
