+++ src/main/java/com/cone/cone/external/socket/constant/SocketIOEvent.java
@@ -1,4 +1,4 @@
-package com.cone.cone.external.socket.type;
+package com.cone.cone.external.socket.constant;
 
 public class SocketIOEvent {
     public static final String MESSAGE = "message";
